[2] add the bootloader
