+++ lab2/bootloader/start.s
@@ -1,0 +1,20 @@
+.section ".text.boot"
+
+.globl _start
+_start:
+    mrs    x0, mpidr_el1        
+    and    x0, x0,#0xFF        // Check processor id
+    cbz    x0, master        // Hang for all non-primary CPU
+
+hang:
+    b hang
+
+master:
+    adr    x0, bss_begin
+    adr    x1, bss_end
+    sub    x1, x1, x0
+    bl     memzero
+
+    mov    sp, #0x400000  // 4MB
+    bl    bootloader_main
+
