+++ lab2/kernel/main.c
