+++ lab2/loadImg/load_img_input
@@ -1,5 +1,3 @@
-
-help
 load_img
 ../kernel8.img
 help
